Fix issue

+++ DataBase/mysql-5.x/UPDATE-a2billing-nixonch.sql
@@ -61,7 +61,7 @@ ALTER TABLE cc_card ADD margin_diller int(3) NOT NULL DEFAULT 10 AFTER `margin`;
 ALTER TABLE cc_card ADD commission DECIMAL( 15, 5 ) NOT NULL DEFAULT 0 AFTER `margin_diller`;
 ALTER TABLE cc_card ADD areaprefix SMALLINT( 6 ) NULL DEFAULT NULL AFTER `country`;
 ALTER TABLE cc_card ADD citylength SMALLINT( 6 ) NULL DEFAULT NULL AFTER `areaprefix`;
-ALTER TABLE cc_card ADD showcallstypedefault TINYINT NOT NULL DEFAULT '0';
+ALTER TABLE cc_card ADD showcallstypedefault INT(11) NOT NULL DEFAULT '0';
 
 ALTER TABLE cc_card CHANGE id_campaign id_campaign  INT( 11 ) NULL DEFAULT '-1';
 ALTER TABLE cc_card CHANGE id_timezone id_timezone CHAR( 40 ) NULL DEFAULT '0';
